requirement.txt
Adafruit-Blinka==8.5.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-mlx90640==1.2.12
adafruit-circuitpython-register==1.9.14
adafruit-circuitpython-requests==1.12.8
adafruit-circuitpython-typing==1.8.2
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.31.0
Adafruit-PureIO==1.1.9
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
firebase-admin==6.0.0
fonttools==4.37.4
gcloud==0.17.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
jws==0.1.3
kiwisolver==1.4.4
matplotlib==3.6.1
msgpack==1.0.4
numpy==1.23.4
oauth2client==3.0.0
packaging==21.3
picamera==1.13
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyftdi==0.54.0
pyparsing==3.0.9
Pyrebase==3.0.27
pyserial==3.5
python-dateutil==2.8.2
python-jwt==2.0.1
pyusb==1.2.1
requests==2.28.1
requests-toolbelt==0.7.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
sysv-ipc==1.1.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
